backend/requirements.txt
Django==4.2.4
django-cors-headers==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Jinja2==3.1.2
requests==2.27.1
Pillow==9.4.0
psycopg2==2.8.6